Refactor: app 모듈의 theme 관련 코드 designsystem 모듈로 이전
