+++ app/src/main/java/com/threegap/bitnagil/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.threegap.bitnagil.ui.theme.BitnagilTheme
+import com.threegap.bitnagil.designsystem.BitnagilTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
